requirements-airflow.txt
# ── Dépendances Airflow ────────────────────────────────────────────────────────
# Installer avec : pip install -r requirements-airflow.txt
#
# ⚠️  Apache Airflow nécessite des contraintes de version pour éviter les conflits.
#     Utilisez la commande d'installation recommandée ci-dessous :
#
#   AIRFLOW_VERSION=2.9.3
#   PYTHON_VERSION=$(python --version | cut -d " " -f 2 | cut -d "." -f 1-2)
#   CONSTRAINT_URL="https://raw.githubusercontent.com/apache/airflow/constraints-${AIRFLOW_VERSION}/constraints-${PYTHON_VERSION}.txt"
#   pip install "apache-airflow==${AIRFLOW_VERSION}" --constraint "${CONSTRAINT_URL}"
#
apache-airflow>=2.9.0

